bug fix: drawing history for updating tool properties
A missing history truncation when accepting a new history step led to
unexpected behavior when changing properties (such as color) of a
selection of items.

Further improvement: If a path has the same color for brush and pen,
then changing the color for this path will affect both brush and pen.
Previously, color changes only affected the pen color.