PixmapGraphicsItem: store pixmaps in simple list
Performance of PixmapGraphicsItem::rendering is highly important.
The old version used a QMap to store pixmaps, which makes it efficient
if a large number of views shows the same slide. This commit replaces
it by an implementation that scales linear in the number of pixmaps,
because usually PixmapGraphicsItem contains only 2 pixmaps are shown.